+++ system/event-script-engine/src/test/java/com/accenture/flows/FlowTests.java
@@ -357,6 +357,8 @@ class FlowTests extends TestBase {
         assertEquals(List.of(Map.of("test", 1), Map.of("test", 1)), body.get("updated_list_map"));
         // f:listToSet() de-duplicates [AAA, AAA] to a list of unique items
         assertEquals(List.of("AAA"), body.get("unique"));
+        // check default value
+        assertEquals("hello", body.get("default"));
     }
 
     @SuppressWarnings("unchecked")